feat: bugfix + 单测

+++ src/main/java/com/github/jiuzhuan/domain/repository/example/domain/entity/OrderGoodDiscountInfo.java
@@ -2,6 +2,8 @@ package com.github.jiuzhuan.domain.repository.example.domain.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author arrety
  * @date 2022/5/14 13:45
@@ -11,6 +13,6 @@ public class OrderGoodDiscountInfo {
 
     public Integer id;
     public Integer slaveOrderInfoId;
-    public Double discount;
+    public BigDecimal discount;
 
 }
